+++ objects/src/accounts/delta/mod.rs
@@ -157,6 +157,18 @@ impl AccountUpdateDetails {
     }
 }
 
+/// Converts an [Account] into an [AccountDelta] for initial delta construction.
+impl From<Account> for AccountDelta {
+    fn from(account: Account) -> Self {
+        let (_id, vault, storage, _code, nonce) = account.into_parts();
+        AccountDelta {
+            storage: storage.into(),
+            vault: (&vault).into(),
+            nonce: Some(nonce),
+        }
+    }
+}
+
 // SERIALIZATION
 // ================================================================================================